Critical fix for all Contexts in the app -- see extended commit message
A re-render of a parent to a context Provider triggers a re-render of the Provider. This in turn caused the context state object to be re-created, which would result in all children re-rendering. This was needless and low-performance (and the cause of the weird "total room re-reload when the sidebar toggled"). I think the correct thing to do is, as-per this commit, `useMemo` all state objects so that if the provider re-renders due to a parent re-render, the state object stays the same (and thus children don't need to re-render).