Shuffle room continuations: Direct people back to the event room (if any)
Instead of mysteriously dumping people back on the general shuffle page,
if there was an event for the current shuffle queue running, take people
back to the room for that event.

This also avoids the need for configuring continuations for the shuffle
queues for this common scenario. This also means it enables the same
shuffle queue to be re-used for multiple events (provided they aren't
overlapping across different rooms).