Shuffle algorithm: Limit room time during events
During a shuffle event people may be allocated to a room after the event started. If the event time was set to the same as the room duration (a common use-case), then joining 60s after the event start would cause you to come back to the event room a minute late.

This change looks at all active events for a shuffle period and reduces the room time to fit the time to the end of the event (if it is shorter than the configured room duration). This ensures people will return to the event room at (near enough) the right time.