Bug fix: Don't activate passive continuations when they're activation threshhold was in the past
This presented itself when moving between backstages - the event just ended was (correctly) still loaded in as the current continuation. The new system for inserting a "next backstage" continuation introduced this situation where the passive continuation choices may be totally re-rendered a second time post-activation (i.e. after moving backstage, since that changes the inputs to the continuation popup when the "next backstage" option is removed)