Enable multi-screen-share; Refactor database handling of vonage layout/participants
The previous model tied vonage session layouts and participants into the EventVonageSession. However, we want to be able to apply layouts to vonage sessions in social rooms / video-chat events (which are handled via the Room's vonage session). Thus these changes split the VonageSessionLayout and VonageParticipantStream into separate tables. These are then tied together with the EventVonageSession and Room tables (which hold vonage session ids for live-stream-events and rooms/video-chat-events respectively).

Introducing a proper VonageSession table to clean up the architecture is a step too far at the moment - too much of the business logic and frontend has grown around the existing system to be able to pull it apart. We'll just wait to hit delete on it when the new video architecture comes around.