Trying to fix up numerous caching glitches caused by Urql query-reload behaviour being different from Apollo
We really need to stop using global contexts that can cause the entire
app to re-render without warning.

We need better handling of this kind of critical stateful logic & its
cache interactions. We're asking for trouble of the kind we saw last
year if we carry on just "hoping" nothing causes one of the 15+ global
contexts to reload the entire app or page.